CannedResponsesListView: Dropdown -> ActionSheet
