+++ app/views/CannedResponsesListView/styles.ts
@@ -7,10 +7,7 @@ export default StyleSheet.create({
 		flex: 1
 	},
 	dropdownContainer: {
-		width: '100%',
-		position: 'absolute',
-		top: 0,
-		borderBottomWidth: StyleSheet.hairlineWidth
+		width: '100%'
 	},
 	backdrop: {
 		...StyleSheet.absoluteFillObject
